refactored into python package, refactored gui into multiple files
