+++ _docs/changelog/index.rst
@@ -8,5 +8,6 @@ Version 0.0
 .. toctree::
     :maxdepth: 2
 
+    v0.0.2
     v0.0.1
     v0.0.0
